# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RankYatra API - Competitive Exam Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get user profile with stats
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload profile picture (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadAvatarBody"
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        walletBalance:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, walletBalance, isAdmin, isBlocked, createdAt]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        walletBalance:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
        totalExamsTaken:
          type: integer
        totalExamsWon:
          type: integer
        bestRank:
          type: ["integer", "null"]
      required: [id, name, email, walletBalance, isAdmin, createdAt, totalExamsTaken, totalExamsWon]
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    UploadAvatarBody:
      type: object
      properties:
        avatarBase64:
          type: string
        mimeType:
          type: string
      required: [avatarBase64, mimeType]
    AvatarResponse:
      type: object
      properties:
        avatarUrl:
          type: string
      required: [avatarUrl]
...